Preview on compiler-optimizable decoder templates
Decoder templates are the next step in generating pre-optimized decoders.
Templates can be written in much cleaner code than the current decoder factory,
which honestly is an unreadable mess and impossible to maintain.
As a bonus, multi-threaded compilation will be doable after removing the
factory, because compilation units aren't required to wait for the factory
anymore to be executed mid-compilation.