Implemented SCAN decoder and fixed garbage output of non-systematic decoders
The currently implemented decoders are all systematic decoders and thus need to
transform the output bits through the butterfly graph in order to get non-
systematic bits. There was a giant bug in nearly every decode()-function,
which produced garbage if asked for non-systematic output.

Following the paper "Low-Complexity Soft-Output Decoding of Polar Codes"
by Fayyaz and Barry, a Soft CANcellation (SCAN)-decoder has been implemented
in this commit, which might lead to corrections in soft-output decoding of
all other decoders implemented so far.
Note on Algorithm 6 in the paper: E and O need to be swapped to make sense
for even phi and odd phi after the increment/decrement.